% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate_ml}
\alias{estimate_ml}
\title{Estimate a Dynamic Factor Model by Maximum Likelihood}
\usage{
estimate_ml(
  dfm,
  data,
  standardized = FALSE,
  input_standardization = NULL,
  pca_init = TRUE,
  em_init = TRUE,
  em_max_iter = 100,
  em_eps = 1e-09,
  max_iter = 1000,
  max_block_iter = 5,
  simpl_model_iter = 15,
  independent_var_shocks = FALSE,
  mixedEstimation = TRUE,
  eps = 1e-09,
  re_estimate = TRUE
)
}
\arguments{
\item{dfm}{An object of class \code{"JD3_DFMMODEL"}, typically generated using the \code{create_model()} function.}

\item{data}{A \code{"mts"} object containing the (transformed) input data.}

\item{standardized}{Boolean. Indicates whether the input data are already standardized.
The default is \code{FALSE}, in which case standardization is applied as a preprocessing step.}

\item{input_standardization}{A matrix specifying the mean and standard deviation of the series to be used during standardization.
The default is \code{NULL}, meaning that these quantities are computed from the data.
This argument can be set using the output of \code{get_results()$preprocessing$sample_mean_stdev} from a previous model estimation. If provided manually, it must be a two-column matrix with the means in the first column and the standard deviations in the second column. The ordering of the rows must match the series in \code{data}.
This argument must be provided if \code{re_estimate = FALSE}, and is ignored if \code{standardized = TRUE}.}

\item{pca_init}{Boolean. Indicates whether a Principal Components Analysis (PCA) is performed beforehand and used to initialize either the EM algorithm (if \code{em_init = TRUE}) or directly the ML estimation. The default is \code{TRUE}.}

\item{em_init}{Boolean. Indicates whether the EM algorithm is run prior to ML estimation and used to provide initial values. The default is \code{TRUE}.}

\item{em_max_iter}{Integer. Specifies the maximum number of iterations for the EM algorithm. Ignored if \code{em_init = FALSE}. The default is \code{100}.}

\item{em_eps}{Numeric. The EM algorithm runs until the increase of the percentage likelihood falls below the \code{eps} value (default is \code{1e-9}), or when the maximum number of iterations is reached. Ignored if \code{em_init = FALSE}.}

\item{max_iter}{Integer. Specifies the maximum number of iterations for the ML estimation. The default is \code{1000}.}

\item{max_block_iter}{Integer. Specifies the maximum number of iterations per block in the optimization process.
Model parameters are divided into two blocks: one for measurement equation and one for VAR equations. Unlike the EM algorithm (one iteration per block), numerical optimization allows multiple iterations per block. The default is \code{5}.}

\item{simpl_model_iter}{Integer. Specifies the number of iterations allowed for the simplified model. The default is \code{15}.}

\item{independent_var_shocks}{Boolean. Indicates whether shocks in the VAR block are assumed to be independent. The default is \code{FALSE}.}

\item{mixedEstimation}{Boolean. Indicates whether to alternate between iterations on the VAR block alone and simultaneous iterations on the two blocks. The default is \code{TRUE}.}

\item{eps}{Numeric. The ML estimation runs until the increase of the percentage likelihood falls below the \code{eps} value (default is \code{1e-9}), or when the maximum number of iterations is reached.}

\item{re_estimate}{Boolean. Indicates whether the model parameters should be re-estimated.
The default is \code{TRUE}. It can be set to \code{FALSE} to keep the model frozen for some period of time, although prolonged use of a frozen model is not recommended.}
}
\value{
An object of class \code{"JD3_DFMESTIMATES"} is returned. The following are
returned invisibly as a list:
\itemize{
\item \code{dfm} \verb{[[1]]} an object of class \code{"JD3_DFMMODEL"} containing the estimated model parameters;
\item \code{data} \verb{[[2]]} the value of the \code{data} argument;
\item \code{is_standardized} \verb{[[3]]} the value of the \code{standardized} argument;
\item \code{input_standardization} \verb{[[4]]} the value of the \code{input_standardization} argument;
\item \code{log_likelihood} \verb{[[5]]} the estimated log-likelihood;
\item \code{gradient} \verb{[[6]]} the estimated gradient at the solution;
\item \code{hessian} \verb{[[7]]} the estimated hessian matrix at the solution;
\item \code{has_converged} \verb{[[8]]} a boolean indicating whether the ML estimation process has converged.
}
}
\description{
Estimate the model parameters by Maximum Likelihood (ML), with initial
values obtained from the Expectations-Maximization (EM) algorithm by default.
The function includes optional arguments to tune the estimation process.
}
\examples{
\dontshow{if (rjd3toolkit::get_java_version() >= rjd3toolkit::minimal_java_version) withAutoprint(\{ # examplesIf}
# input data
set.seed(100)
data <- ts(matrix(rnorm(500), 100, 5),
           frequency = 12,
           start = c(2010, 1))
data[100, 1] <- data[99:100, 2] <- data[(1:100)[-seq(3, 100, 3)], 5] <- NA

# define a new model
dfm <- create_model(
    nfactors = 2,
    nlags = 2,
    factors_type = c("M", "M", "YoY", "M", "Q"),
    factors_loading = matrix(data = TRUE, 5, 2),
    var_init = "Unconditional"
)

# estimate by maximum likelihood
est_ml_a <- estimate_ml(dfm, data)

# results and forecasts
rslts_a <- get_results(est_ml_a)
fcsts_a <- get_forecasts(est_ml_a)

# no re-estimation of a previous model while integrating updated data in the output object
data_new <- data
data_new[99, 2] <- 1
est_ml_b <- estimate_ml(est_ml_a$dfm,
                        data_new,
                        input_standardization = rslts_a$preprocessing$sample_mean_stdev,
                        re_estimate = FALSE)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{create_model()} to define a new model,

\code{estimate_pca()} for estimation using principal components analysis,

\code{estimate_em()} for estimation using EM algorithm,

\code{get_results()} to access estimation results,

\code{get_forecasts()} to obtain forecasts.

For more information, see the vignette:

\code{utils::browseVignettes()}, e.g. \code{browseVignettes(package = "rjd3nowcasting")}
}
