% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{get_residuals}
\alias{get_residuals}
\title{Get residuals of standardized residuals}
\usage{
get_residuals(results)
}
\arguments{
\item{results}{an object of class 'JD3_SsfModelEstimation'}
}
\value{
a list
}
\description{
Residuals are the one-step forecast errors, while the standardized residuals
are the one-step forecast errors after standardization (i.e. divided by
their standard deviation). The basic diagnostics for normality,
heteroscedasticity and auto-correlation should be performed on the
standardized residuals.
}
\examples{
set.seed(100)
data<-ts(matrix(rnorm(500), 100, 5), frequency = 12, start = c(2010,1))
data[100,1]<-data[99:100,2]<-data[(1:100)[-seq(3,100,3)],5]<-NA
dfm_model <- model(nfactors=2,
                   nlags=2,
                   factors_type = c("M", "M", "YoY", "M", "Q"),
                   factors_loading = matrix(data=TRUE, 5, 2),
                   var_init = "Unconditional")
rslt_em<-estimate_em(dfm_model, data)
get_residuals(rslt_em)

}
