% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{macroIndicators}
\alias{macroIndicators}
\alias{data0}
\alias{data1}
\title{Datasets including some French macro-economic variables}
\format{
An object of class \code{data.frame} with 150 rows and 11 columns.

An object of class \code{data.frame} with 150 rows and 11 columns.
}
\usage{
data0

data1
}
\description{
The datasets 'data0' and 'data1' acts as successive releases of
macro-economic time series. They contain data on monthly industrial
production index (PVI), turnover (TURN), quarterly GDP, as well as
business survey data (BS) and other survey data (PMI) for both France
and the Eurozone. Those datasets are used to illustrate how one of these
variable can be nowcasted using the others using a Dynamic Factor model.
}
\keyword{datasets}
