% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{model}
\alias{model}
\title{Create Dynamic Factor Model}
\usage{
model(
  nfactors,
  nlags,
  factors_type,
  factors_loading,
  var_init = c("Unconditional", "Zero"),
  mVariance = NULL
)
}
\arguments{
\item{nfactors}{Integer. Number of factors.}

\item{nlags}{Integer. Number of lags in VAR.}

\item{factors_type}{Character vector. Respecting the order of the series in
 the input data, you must refer here the link between the (transformed)
 series and the factors. Three options are possible:
\itemize{
  \item{"M"}{Variables expressed in terms of monthly growth rates can be
  linked to a factor representing the underlying monthly growth rate of the
  economy if "M" is selected}
  \item{"Q"}{Monthly or quarterly variables that are correlated with the the
  underlying quarterly growth rate of the economy can be linked to a
  weighted average of the factors representing the underlying monthly
  growth rate of the economy. Such a weighted average is meant to represent
  quarterly growth rates, and it is implemented by selecting "Q"}
  \item{"YoY"}{The variables can also be linked to the cumulative sum of the
  last 12 monthly factors. If the model is designed in such a way that the
  monthly factors represent monthly growth rates, the resulting cumulative
  sum boils down to the year-on-year growth rate. Thus, variables expressed
  in terms of year-on-year growth rates or surveys that are correlated with
  the year-on-year growth rates of the reference series should be linked to
  the factors using "YoY".}
}}

\item{factors_loading}{Boolean matrix. It represents the factor loading
structure. The dimension of the matrix should be 'number of series' x
'number of factors'. For each row representing each series, the user must
mention whether the corresponding factor load on this series.}

\item{var_init}{Character. The first unobserved factors values in the sample
is assumed either to be equal to zero or be consistent with a normal
distribution with mean zero and a variance consistent with the
unconditional variance of the VAR. The latter is the default.}

\item{mVariance}{Not yet implemented. Keep it on NULL.}
}
\value{
an object of class 'JD3_DfmModel'
}
\description{
Create Dynamic Factor Model
}
\examples{
dfm_model <- model(nfactors=2,
                   nlags=2,
                   factors_type = c("M", "M", "YoY", "M", "Q"),
                   factors_loading = matrix(data=TRUE, 5, 2),
                   var_init = "Unconditional")

}
