% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot.JD3_DfmForecasts}
\alias{plot.JD3_DfmForecasts}
\title{Plot function for objects of class 'JD3_DfmForecasts'}
\usage{
\method{plot}{JD3_DfmForecasts}(x, series_name = NULL, ...)
}
\arguments{
\item{x}{an object of class 'JD3_DfmForecasts'}

\item{series_name}{Character. Name of the series to plot. By default, the
first series will be plotted.}

\item{\dots}{further arguments passed to ts.plot().}
}
\description{
Plot function for objects of class 'JD3_DfmForecasts'
}
