% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.R
\name{plot.JD3_DfmNews}
\alias{plot.JD3_DfmNews}
\title{Plot function for objects of class 'JD3_DfmNews'}
\usage{
\method{plot}{JD3_DfmNews}(x, ...)
}
\arguments{
\item{x}{an object of class 'JD3_DfmNews'}

\item{\dots}{further arguments passed to barplot().}
}
\description{
Plot function for objects of class 'JD3_DfmNews'
}
