% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot.JD3_SsfModelEstimation}
\alias{plot.JD3_SsfModelEstimation}
\title{Plot function for objects of class 'JD3_SsfModelEstimation'}
\usage{
\method{plot}{JD3_SsfModelEstimation}(x, series_name = NULL)
}
\arguments{
\item{x}{an object of class 'JD3_SsfModelEstimation'}

\item{series_name}{Character. Name of the series to plot. By default, the
first series will be plotted.}
}
\description{
Plot function for objects of class 'JD3_SsfModelEstimation'
}
