% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{print.JD3_DfmEstimates}
\alias{print.JD3_DfmEstimates}
\title{Print function for objects of class 'JD3_DfmEstimates'}
\usage{
\method{print}{JD3_DfmEstimates}(x, ...)
}
\arguments{
\item{x}{an object of class 'JD3_DfmEstimates'}

\item{\dots}{further arguments passed to the print() function.}
}
\description{
Print function for objects of class 'JD3_DfmEstimates'
}
