% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{print.JD3_DfmForecasts}
\alias{print.JD3_DfmForecasts}
\title{Print function for objects of class 'JD3_DfmForecasts'}
\usage{
\method{print}{JD3_DfmForecasts}(x, ...)
}
\arguments{
\item{x}{an object of class 'JD3_DfmForecasts'}

\item{\dots}{further arguments passed to the print() function.}
}
\description{
Print function for objects of class 'JD3_DfmForecasts'
}
