% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.R
\name{print.JD3_DfmNews}
\alias{print.JD3_DfmNews}
\title{Print function for objects of class 'JD3_DfmNews'}
\usage{
\method{print}{JD3_DfmNews}(x, ...)
}
\arguments{
\item{x}{an object of class 'JD3_DfmNews'}

\item{\dots}{further arguments passed to the print() function.}
}
\description{
Print function for objects of class 'JD3_DfmNews'
}
