% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{print.JD3_DfmResults}
\alias{print.JD3_DfmResults}
\title{Print function for objects of class 'JD3_DfmResults'}
\usage{
\method{print}{JD3_DfmResults}(x, ...)
}
\arguments{
\item{x}{an object of class 'JD3_DfmResults'}

\item{\dots}{further arguments passed to the print() function.}
}
\description{
Print function for objects of class 'JD3_DfmResults'
}
