% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{print.JD3_SsfModelEstimation}
\alias{print.JD3_SsfModelEstimation}
\title{Print function for objects of class 'JD3_SsfModelEstimation'}
\usage{
\method{print}{JD3_SsfModelEstimation}(x)
}
\arguments{
\item{x}{an object of class 'JD3_SsfModelEstimation'}
}
\description{
Print function for objects of class 'JD3_SsfModelEstimation'
}
