% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.R
\name{summary.JD3_DfmNews}
\alias{summary.JD3_DfmNews}
\title{Summary function for objects of class 'JD3_DfmNews'}
\usage{
\method{summary}{JD3_DfmNews}(object, ...)
}
\arguments{
\item{object}{an object of class 'JD3_DfmNews'}

\item{\dots}{further arguments passed to the print() function.}
}
\description{
Summary function for objects of class 'JD3_DfmNews'
}
