% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{summary.JD3_SsfModelEstimation}
\alias{summary.JD3_SsfModelEstimation}
\title{Summary function for objects of class 'JD3_SsfModelEstimation'}
\usage{
\method{summary}{JD3_SsfModelEstimation}(x)
}
\arguments{
\item{x}{an object of class 'JD3_SsfModelEstimation'}
}
\description{
Summary function for objects of class 'JD3_SsfModelEstimation'
}
