% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{create_model}
\alias{create_model}
\title{Create a Dynamic Factor Model}
\usage{
create_model(
  nfactors,
  nlags,
  factors_type,
  factors_loading,
  var_init = c("Unconditional", "Zero"),
  var_coefficients = NULL,
  var_errors_variance = NULL,
  measurement_coefficients = NULL,
  measurement_errors_variance = NULL
)
}
\arguments{
\item{nfactors}{Integer. Number of factors.}

\item{nlags}{Integer. Number of lags in the VAR process.}

\item{factors_type}{Character vector. Defines, in the order of the input series, how each (transformed) series relates to the factors.
Available options are:
\itemize{
\item{"M"}{: Variables expressed as monthly growth rates, linked directly to the latent factors.}
\item{"Q"}{: Quarterly or monthly variables linked to a weighted average of the factors, representing quarterly growth rates.}
\item{"YoY"}{: Variables linked to the cumulative sum of the last 12 monthly factors, corresponding to year-on-year growth rates. This option is appropriate for variables expressed in year-on-year terms, or for series that are closely related to such evolution.}
}}

\item{factors_loading}{Boolean matrix defining the factor loading structure, with dimension \emph{number of series} × \emph{number of factors}. Each entry indicates whether a given factor loads on a given series.}

\item{var_init}{Character. Specifies the initialization of the latent factors: either zero or assumed to follow a normal distribution with mean zero and variance equal to the unconditional variance of the VAR (default).}

\item{var_coefficients}{Matrix. The VAR coefficients. If \code{NULL} (default), they will be estimated from scratch.
Otherwise, user-defined values can be provided as starting point for the estimation, typically obtained from a previous estimation of the model.}

\item{var_errors_variance}{Matrix. The variance-covariance matrix of the VAR errors. If \code{NULL} (default), it will be estimated from scratch.
Otherwise, user-defined values can be provided as starting point for the estimation, typically obtained from a previous estimation of the model.}

\item{measurement_coefficients}{Matrix. The measurement equation coefficients. If \code{NULL} (default), they will be estimated from scratch.
Otherwise, user-defined values can be provided as starting point for the estimation, typically obtained from a previous estimation of the model.}

\item{measurement_errors_variance}{Numeric vector. The variance of the idiosyncratic measurement errors. If \code{NULL} (default), they will be estimated from scratch.
Otherwise, user-defined values can be provided as starting point for the estimation, typically obtained from a previous estimation of the model.}
}
\value{
An object of class \code{"JD3_DFMMODEL"} is returned. The following are
returned invisibly as a list:
\itemize{
\item \code{var_coefficients} \verb{[[1]]} initial values of the VAR coefficients;
\item \code{var_errors_variance} \verb{[[2]]} initial values of the variance-covariance matrix of the VAR errors;
\item \code{measurement_coefficients} \verb{[[3]]} initial values of the measurement equation coefficients;
\item \code{measurement_errors_variance} \verb{[[4]]} initial values of the variance of the idiosyncratic measurement errors;
\item \code{initialization_type} \verb{[[5]]} the value of the \code{var_init} argument;
\item \code{factors_type} \verb{[[6]]} the value of the \code{factors_type} argument.
}
}
\description{
Creates a Dynamic Factor Model (DFM) by specifying the number of factors and
the number of lags in the Vector Auto-Regressive (VAR) process. The
function also allows the user to choose between different types of links
between each series and the latent factors, in addition to defining the
factor loading structure.
}
\examples{
\dontshow{if (rjd3jars::check_java_version(silent = TRUE)) withAutoprint(\{ # examplesIf}
# From scratch
dfm1 <- create_model(nfactors = 2,
                     nlags = 2,
                     factors_type = c("M", "M", "YoY", "M", "Q"),
                     factors_loading = matrix(data = TRUE, 5, 2),
                     var_init = "Unconditional")

# From a previously estimated model
set.seed(100)
data <- ts(matrix(rnorm(500), 100, 5),
           frequency = 12,
           start = c(2010, 1))
data[100, 1] <- data[99:100, 2] <- data[(1:100)[-seq(3, 100, 3)], 5] <- NA
est1 <- estimate_em(dfm1, data)
dfm2 <- create_model(nfactors = 2,
                     nlags = 2,
                     factors_type = c("M", "M", "YoY", "M", "Q"),
                     factors_loading = matrix(data = TRUE, 5, 2),
                     var_init = "Unconditional",
                     var_coefficients = est1$dfm$var_coefficients,
                     var_errors_variance = est1$dfm$var_errors_variance,
                     measurement_coefficients = est1$dfm$measurement_coefficients,
                     measurement_errors_variance = est1$dfm$measurement_errors_variance)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{estimate_pca()} to estimate the generated model using principal components analysis,

\code{estimate_em()} to estimate the generated model using the Expectations-Maximization algorithm,

\code{estimate_ml()} to estimate the generated model using maximum likelihood.

For more information, see the vignette:

\code{utils::browseVignettes()}, e.g. \code{browseVignettes(package = "rjd3nowcasting")}
}
