% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate_em}
\alias{estimate_em}
\title{Estimate DFM with Expectations-Maximization algorithm}
\usage{
estimate_em(
  dfmModel,
  data,
  standardized = FALSE,
  n_out = 12,
  pca_init = TRUE,
  max_iter = 100,
  eps = 1e-09
)
}
\arguments{
\item{dfmModel}{an object of class 'JD3_DfmModel'. Typically generated by the
model() function.}

\item{data}{an mts object.}

\item{standardized}{Boolean. Indicate whether the input series were already
standardized or not. Default is FALSE, i.e. not previously standardized}

\item{n_out}{Number of out-of-periods considered in the processing. Must be
greater than the required number of forecasts asked in the forecasting
funcions.}

\item{pca_init}{Boolean. Indicate whether a principal components analysis
is performed beforehand and used as an initial condition for the EM
algorithm.}

\item{max_iter}{Integer. Maximum number of iterations of the EM algorithm.}

\item{eps}{Numeric. EM algorithm is run until the percentage likelihood does
not increase by more than the eps value (1e-9 is the default) or until the
maximum number of iterations is hit.}
}
\value{
an object of class 'JD3_SsfModelEstimation'
}
\description{
Estimate DFM with Expectations-Maximization algorithm
}
\examples{
set.seed(100)
data<-ts(matrix(rnorm(500), 100, 5), frequency = 12, start = c(2010,1))
data[100,1]<-data[99:100,2]<-data[(1:100)[-seq(3,100,3)],5]<-NA
dfm_model <- model(nfactors=2,
                   nlags=2,
                   factors_type = c("M", "M", "YoY", "M", "Q"),
                   factors_loading = matrix(data=TRUE, 5, 2),
                   var_init = "Unconditional")
rslt_em<-estimate_em(dfm_model, data)

}
