% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate_ml}
\alias{estimate_ml}
\title{Estimate DFM with Maximum Likelihood}
\usage{
estimate_ml(
  dfmModel,
  data,
  standardized = FALSE,
  n_out = 12,
  pca_init = TRUE,
  em_init = TRUE,
  em_max_iter = 100,
  em_eps = 1e-09,
  max_iter = 1000,
  max_block_iter = 5,
  simpl_model_iter = 15,
  independent_var_shocks = FALSE,
  mixedEstimation = TRUE,
  eps = 1e-09
)
}
\arguments{
\item{dfmModel}{an object of class 'JD3_DfmModel'. Typically generated by the
model() function.}

\item{data}{an mts object.}

\item{standardized}{Boolean. Indicate whether the input series were already
standardized or not. Default is FALSE, i.e. not previously standardized}

\item{n_out}{Number of out-of-periods considered in the processing. Must be
greater than the required number of forecasts asked in the forecasting
funcions.}

\item{pca_init}{Boolean. Indicate whether a principal components analysis is
performed beforehand and used as an initial condition for the EM algorithm
(if em_init=TRUE) or directly for the estimation by ML.}

\item{em_init}{Boolean. Indicate whether the EM algorithm is performed
beforehand and used as an initial condition for the estimation by ML.}

\item{em_max_iter}{Integer. Maximum number of iterations of the EM algorithm.
Ignored if em_init = FALSE.}

\item{em_eps}{Numeric. EM algorithm is run until the percentage likelihood
does not increase by more than the eps value (1e-9 is the default) or until
the maximum number of iterations is hit. Ignored if em_init = FALSE.}

\item{max_iter}{Integer. Maximum number of iterations for the ML estimation.}

\item{max_block_iter}{Integer. Maximum number of iterations in optimization
by block. The model parameters are divided in two blocks: one related to
the measurement equations and one to the VAR. While the EM algorithm
requires one iteration per block, the numerical optimization allows us to
set the number of iterations desired per block.}

\item{simpl_model_iter}{Integer. Number of simplified model iterations
allowed.}

\item{independent_var_shocks}{Boolean. Whether we assume that shocks in the
VAR block are independent.}

\item{mixedEstimation}{Boolean. The mixed estimation option alternates
between the iterations for the VAR block alone and simultaneous iterations
for the two blocks.}

\item{eps}{Numeric. ML estimation is run until the percentage likelihood does
not increase by more than the eps value (1e-9 is the default) or until the
maximum number of iterations is hit.}
}
\value{
an object of class 'JD3_SsfModelEstimation'
}
\description{
Estimate DFM with Maximum Likelihood
}
\examples{
set.seed(100)
data<-ts(matrix(rnorm(500), 100, 5), frequency = 12, start = c(2010,1))
data[100,1]<-data[99:100,2]<-data[(1:100)[-seq(3,100,3)],5]<-NA
dfm_model <- model(nfactors=2,
                   nlags=2,
                   factors_type = c("M", "M", "YoY", "M", "Q"),
                   factors_loading = matrix(data=TRUE, 5, 2),
                   var_init = "Unconditional")
rslt_ml<-estimate_ml(dfm_model, data)

}
