% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate_pca}
\alias{estimate_pca}
\title{Estimate a Dynamic Factor Model Using Principal Components Analysis}
\usage{
estimate_pca(
  dfm,
  data,
  standardized = FALSE,
  input_standardization = NULL,
  re_estimate = TRUE
)
}
\arguments{
\item{dfm}{An object of class \code{"JD3_DFMMODEL"}, typically generated using the \code{create_model()} function.}

\item{data}{A \code{"mts"} object containing the (transformed) input data.}

\item{standardized}{Boolean. Indicates whether the input data are already standardized.
The default is \code{FALSE}, in which case standardization is applied as a preprocessing step.}

\item{input_standardization}{A matrix specifying the mean and standard deviation of the series to be used during standardization.
The default is \code{NULL}, meaning that these quantities are computed from the data.
This argument can be set using the output of \code{get_results()$preprocessing$sample_mean_stdev} from a previous model estimation. If provided manually, it must be a two-column matrix with the means in the first column and the standard deviations in the second column. The ordering of the rows must match the series in \code{data}.
This argument must be provided if \code{re_estimate = FALSE}, and is ignored if \code{standardized = TRUE}.}

\item{re_estimate}{Boolean. Indicates whether the model parameters should be re-estimated.
The default is \code{TRUE}. It can be set to \code{FALSE} to keep the model frozen for some period of time, although prolonged use of a frozen model is not recommended.}
}
\value{
An object of class \code{"JD3_DFMESTIMATES"} is returned. The following are
returned invisibly as a list:
\itemize{
\item \code{dfm} \verb{[[1]]} an object of class \code{"JD3_DFMMODEL"} containing the estimated model parameters;
\item \code{data} \verb{[[2]]} the value of the \code{data} argument;
\item \code{is_standardized} \verb{[[3]]} the value of the \code{standardized} argument;
\item \code{input_standardization} \verb{[[4]]} the value of the \code{input_standardization} argument;
\item additional elements not relevant for Principal Components Analysis estimates.
}
}
\description{
Estimate the model parameters using Principal Component Analysis (PCA).
While this approach is fast, it is generally not recommended to rely on it
exclusively, particularly when the dataset includes quarterly series or
variables related to year-on-year growth rates.
}
\examples{
\dontshow{if (rjd3jars::check_java_version(silent = TRUE)) withAutoprint(\{ # examplesIf}
# input data
set.seed(100)
data <- ts(matrix(rnorm(500), 100, 5),
           frequency = 12,
           start = c(2010, 1))
data[100, 1] <- data[99:100, 2] <- data[(1:100)[-seq(3, 100, 3)], 5] <- NA

# define a new model
dfm <- create_model(
    nfactors = 2,
    nlags = 2,
    factors_type = c("M", "M", "M", "M", "M"),
    factors_loading = matrix(data = TRUE, 5, 2),
    var_init = "Unconditional"
)

# estimate using PCA
est_pca_a <- estimate_pca(dfm, data)

# results and forecasts
rslts_a <- get_results(est_pca_a)
fcsts_a <- get_forecasts(est_pca_a)

# no re-estimation of a previous model while integrating updated data in the output object
data_new <- data
data_new[99, 2] <- 1
est_pca_b <- estimate_pca(est_pca_a$dfm,
                          data_new,
                          input_standardization = rslts_a$preprocessing$sample_mean_stdev,
                          re_estimate = FALSE)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{create_model()} to define a new model,

\code{estimate_em()} for estimation using the Expectations-Maximization algorithm,

\code{estimate_ml()} for estimation using maximum likelihood,

\code{get_results()} to access estimation results,

\code{get_forecasts()} to obtain forecasts.

For more information, see the vignette:

\code{utils::browseVignettes()}, e.g. \code{browseVignettes(package = "rjd3nowcasting")}
}
