% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate_pca}
\alias{estimate_pca}
\title{Estimate DFM with Principal components Analysis}
\usage{
estimate_pca(dfmModel, data, standardized = FALSE, n_out = 12)
}
\arguments{
\item{dfmModel}{an object of class 'JD3_DfmModel'. Typically generated by the
model() function.}

\item{data}{an mts object.}

\item{standardized}{Boolean. Indicate whether the input series were already
standardized or not. Default is FALSE, i.e. not previously standardized}

\item{n_out}{Number of out-of-periods considered in the processing. Must be
greater than the required number of forecasts asked in the forecasting
funcions.}
}
\value{
an object of class 'JD3_SsfModelEstimation'
}
\description{
Estimate DFM with Principal components Analysis
}
\examples{
set.seed(100)
data <- ts(matrix(rnorm(500), 100, 5), frequency = 12, start = c(2010, 1))
data[100, 1] <- data[99:100, 2] <- data[(1:100)[-seq(3, 100, 3)], 5] <- NA
dfm_model <- model(
    nfactors = 2,
    nlags = 2,
    factors_type = c("M", "M", "YoY", "M", "Q"),
    factors_loading = matrix(data = TRUE, 5, 2),
    var_init = "Unconditional"
)
rslt_pca <- estimate_pca(dfm_model, data)

}
