% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.R
\name{get_news}
\alias{get_news}
\title{Perform News Analysis for Forecast Updates in Dynamic Factor Models}
\usage{
get_news(dfm_estimates, new_data, target_series = NULL, n_fcst = 3)
}
\arguments{
\item{dfm_estimates}{An object of class \code{"JD3_DFMESTIMATES"}, typically generated using the \code{estimate_ml()}, \code{estimate_em()}, or \code{estimate_pca()} function.}

\item{new_data}{A \code{"mts"} object containing the updated dataset.}

\item{target_series}{Character. Name of the target series of interest. By default, the first series is used.}

\item{n_fcst}{Integer. Number of forecast periods to consider. The default is \code{3}.}
}
\value{
An object of class \code{"JD3_DFMNEWS"} is returned. The following are
returned invisibly as a list:
\itemize{
\item \code{target_series} \verb{[[1]]} the value of the \code{target_series} argument;
\item \code{weights_T} \verb{[[2]]} the weights of the news for the transformed series;
\item \code{impacts_T} \verb{[[3]]} the impacts of the news (weights x news) for the transformed series;
\item \code{forecasts_T} \verb{[[4]]} the old, revised (including past data revisions) and new forecasts of the transformed series;
\item \code{weights} \verb{[[5]]} the weights of the news for the original series;
\item \code{impacts} \verb{[[6]]} the impacts of the news (weights x news) for the original series;
\item \code{forecasts} \verb{[[7]]} the old, revised (including past data revisions) and new forecasts of the original series.
}
}
\description{
Performs a news analysis to quantify the impact of newly released data on
forecast revisions between two consecutive dataset updates. The function
isolates the contribution of new observations by comparing them to their
forecasts obtained from the revised dataset (including past data revisions)
and provides a detailed decomposition of the resulting forecast changes.
}
\examples{
\dontshow{if (rjd3toolkit::get_java_version() >= rjd3toolkit::minimal_java_version) withAutoprint(\{ # examplesIf}
set.seed(100)
data_t1 <- ts(matrix(rnorm(500), 100, 5),
              frequency = 12,
              start = c(2010, 1))
data_t1[100, 1] <- data_t1[99:100, 2] <- data_t1[(1:100)[-seq(3, 100, 3)], 5] <- NA
data_t2 <- ts(rbind(data_t1, rep(NA, 5)),
              frequency = 12,
              start = c(2010, 1))
data_t2[100, 1] <- data_t2[99, 2] <- data_t2[101, 3] <- data_t2[101, 4] <- 1

dfm_model <- create_model(
    nfactors = 2,
    nlags = 2,
    factors_type = c("M", "M", "YoY", "M", "Q"),
    factors_loading = matrix(TRUE, 5, 2),
    var_init = "Unconditional"
)

est_em_a <- estimate_em(dfm_model, data_t1)

news_a <- get_news(est_em_a, data_t2, target_series = "Series 2", n_fcst = 2)
\dontshow{\}) # examplesIf}
}
\references{
Banbura, M., & Modugno, M. (2010). Maximum likelihood estimation of factor
models on data sets with arbitrary patterns of missing data.
}
\seealso{
\code{create_model()} to define a new Dynamic Factor model,

\code{estimate_pca()} for estimation using principal components analysis,

\code{estimate_em()} for estimation using EM algorithm,

\code{estimate_ml()} for estimation using maximum likelihood,

For more information, see the vignette:

\code{utils::browseVignettes()}, e.g. \code{browseVignettes(package = "rjd3nowcasting")}
}
